Add FLOAT and DOUBLE data interchange functions
In most situations, printing floats and doubles is good enough.
However, there are rare times when we need to do a binary comparison
or exact binary transfer of floating-point values. For example, the
Java programming language provides a number of functions to address
these needs, for instance Float.floatToIntBits and the matching
Float.intBitsToFloat.

This commit adds the following two pairs of functions which allow
single or double precision floating point values to be represented
as integer values and converted back to the same exact floating
point, bit-for-bit.

	float_to_int32_bits(expr)
	int32_bits_to_float(expr)

	double_to_int64_bits(expr)
	int64_bits_to_double(expr)

Example usage:

	MariaDB [test]> SELECT float_to_int32_bits(-1234.5);
	+------------------------------+
	| float_to_int32_bits(-1234.5) |
	+------------------------------+
	|                   -996519936 |
	+------------------------------+
	1 row in set (0.000 sec)

	MariaDB [test]> SELECT int32_bits_to_float(-996519936);
	+---------------------------------+
	| int32_bits_to_float(-996519936) |
	+---------------------------------+
	|                         -1234.5 |
	+---------------------------------+
	1 row in set (0.000 sec)

	MariaDB [test]>

Underlying implementation:

IEEE 754 describes interchange formats, which in the case of our
radix-2, 32-bit FLOAT and 64-bit DOUBLE type, we can memcpy to a
int32 or to an int64.

For information on floating point data exchange formats, see:
https://en.wikipedia.org/wiki/IEEE_754-1985
https://en.wikipedia.org/wiki/IEEE_754-2008

See also:
https://en.wikipedia.org/wiki/Floating-point_arithmetic#IEEE_754:_floating_point_in_modern_computers
https://docs.oracle.com/javase/7/docs/api/java/lang/Float.html#floatToIntBits(float)
https://docs.oracle.com/javase/7/docs/api/java/lang/Float.html#intBitsToFloat(int)
https://docs.oracle.com/javase/7/docs/api/java/lang/Double.html#doubleToLongBits(double)
https://docs.oracle.com/javase/7/docs/api/java/lang/Double.html#longBitsToDouble(long)

I'd like to thank Alexander Barkov for many great pieces of advice,
including argument checking, behavior of the API, asking about
corner cases, and more. And wlad on IRC gave a lot of good advice.
And svoj for helping me understand decimal FIXED and NOT_FIXED, and
his review work.