WIP: MDEV-26877: allow deeply nested vaid JSON values
WIP: `mtr --suite=json` passes!
WIP: However, `func_json` is not in that suite and it never completes.
WIP: Thus, we need to sort out error with:  mtr --do-test=func_json

The json_engine.stack and json_path.steps are now able to grow if
needed. They still start as stack allocated, but if they need to grow
beyond JSON_DEPTH_LIMIT, then they do so.

Function json_find_path no longer requires the array_counters parameter.
The array_counters are scratch-space for the internal workings of the
function. In every usage, the array is stack-allocated, and not reused,
thus there little reason to ask the caller to do this.  This moves the
stack declaration inside the function, and changes it to be a
DYNAMIC_ARRAY allowing it to grow larger than JSON_DEPTH_LIMIT should
that be needed.

Includes a unit test for a JSON object with very deep nesting, and
updates the acceptance test.

Signed-off-by: Eric Herman <eric@freesa.org>